Change submit cli to batch submit cli
