+++ cli/sawtooth_cli/main.py
@@ -41,8 +41,6 @@ from sawtooth_cli.state import add_state_parser
 from sawtooth_cli.state import do_state
 from sawtooth_cli.cluster import add_cluster_parser
 from sawtooth_cli.cluster import do_cluster
-from sawtooth_cli.submit import add_submit_parser
-from sawtooth_cli.submit import do_submit
 
 
 def create_console_handler(verbose_level):
@@ -105,7 +103,6 @@ def create_parser(prog_name):
     add_transaction_parser(subparsers, parent_parser)
     add_state_parser(subparsers, parent_parser)
     add_cluster_parser(subparsers, parent_parser)
-    add_submit_parser(subparsers, parent_parser)
 
     return parser
 
@@ -138,8 +135,6 @@ def main(prog_name=os.path.basename(sys.argv[0]), args=sys.argv[1:],
         do_state(args)
     elif args.command == 'cluster':
         do_cluster(args)
-    elif args.command == 'submit':
-        do_submit(args)
     else:
         raise AssertionError("invalid command: {}".format(args.command))
